pyproject.toml
[tool.poetry]
name = "power-of-words"
version = "0.1.0"
description = ""
authors = ["Zoltan Pal <palzoltanet@gmail.com>"]
readme = "README.md"
# packages = [{include = "pow"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
sqlalchemy = "^2.0.29"
flask-cors = "^4.0.0"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
python-dateutil = "^2.9.0.post0"
feedparser = "^6.0.11"
nltk = "^3.8.1"
transformers = "^4.40.0"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
torch = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
